Allow clicking on options on Safari
Safari triggers focusOut before click, but if you
preventDefault on mouseDown, you can stop that from happening.
If this is removed, clicking on an option in Safari will
trigger `handleOptionFocusOut`, which closes the menu, and the click
will trigger on the element underneath instead.
See: http://stackoverflow.com/questions/7621711/how-to-prevent-blur-running-when-clicking-a-link-in-jquery