Use UMD module definition
This will allow the module to be used both standalone in browser (as per the README documentation) but also compiled into an application bundle when installed into a project from npm (i.e. using browserify or webpack in a parent project).

Note the `module.exports` in wrapper.jsx - it would be probably preferable for this to be `export default`, but this results in the module being wrapped in a `{ __esmodule: true, default: ... }` construct by webpack, which makes a mess of the API. By using `module.exports` instead then the exported module is a direct reference to `AccessibleTypeahead`.