Update aria status region to use , or ,,
This seems like a simpler way to force it to update compared to the
debounced setTimeout. It has the side effect of making NVDA pronounce
"No search results dot."

This allows us to remove the dependency on lodash.debounce.

Also add the aria-atomic attribute which works well in JAWS and NVDA
to get them to reiterate the whole content of the region when
something changes.