Allow an empty `defaultValue` option for `enhanceSelectElement`
The check for a default value option matches false-y values, but that also catches the case where the default value is explicitly set to an empty string.

Add an extra condition for the case where the default value is '' so that it is not overwritten with the content of the first `<option>`.