Do not copy select element name attribute to autocomplete input
Copying the name attribute means that the same `name` fields is POSTed twice on form submission. This can cause server-side parsers to fail because they receive two fields with the same key. By not copying the name attribute onto the autocomplete element we can ensure that only one value is POSTed to the server on form submission.