Fix onChange/onInput event handler
I naively changed this from an `onInput` to an `onChange` when I saw
that it "worked fine" in Preact. However, it wasn't working in
reality; the `pollInputElement` handler was making it look like it.

We can't use `onInput` in both Preact and React because it fails in
React on IE10.