fix: selection and timeout race condition
Fixes issue where if handleOptionClick is triggered and before rendering getDirectInputChanges is also triggered by the polling that pollInputElement causing the typeahead to not properly select the selected item. Caused by the state being updated to the selected item and menu to close however getDirectInputChanges will then change the state back so the item looks to not be selected and menu remains open.