Update wdio-selenium-standalone-service and webdriverio (fixes failing tests)
This updates wdio-selenium-standalone-service from 0.0.8 to 0.0.10,
and webdriverio from 4.8.0 to 4.13.0.

Running `npm outdated` in fact does not require
wdio-selenium-standalone-service to be updated but WebDriverIO is not able
to run tests in Chrome unless wdio-selenium-standalone-service is v0.0.10.
This is a known issue: https://github.com/webdriverio/webdriverio/issues/2631

webdriverio updates include bugfixes and new features.

wdio-selenium-standalone-service by webdriverio does not publish a changelog.