Update standard from 10.0.2 to 11.0.1
Running `npm outdated` in fact only requires
standard to be updated to 10.0.3 but but this caused tests to fail
due to static properties being picked up linter as not defined,
as reported here:
https://github.com/ricardofbarros/linter-js-standard/issues/113

The main change for 11.0.1 is updating `eslint` to version 4 with
most changes being to do with indentation rules.

This commit fixes some indentation issues now picked up by linter in
 `src/autocomplete.js`

Tested to work by running all tests and ensuring that the review app
runs with working examples.