Always debounce rendered aria live content. Improves rapid typing behaviour on IE + JAWS, when virtual PC cursor mode is enabled. Also appears to improve the stale result list problem that sometimes occurs in NVDA.
Remove safari browser detection entirely - in favour of a longer debounce time in all scenarios. Improves the feel of the aria live announcement generally as it gives the user time to type, and also has the side effect of allowing the VoiceOver typing echo to complete before announcing the update.