Remove retries for functional tests
The wdio-sauce-service, which is responsible for reporting the build state back to Saucelabs, does not seem to take into account retries – we believe that if a test fails, it will mark the build as failed in Saucelabs even if it is retried and passes. This is confusing, as it does not match what we see in the Mocha reporter either locally or in the Travis build logs.

Using retries also masks situations where for example the autocomplete state is not reset correctly between tests, and thus the first attempt of a test always fails.

For these reasons, we're going to try removing retries and see what problems it causes (or, to look at another way, unmasks)