Reload the page between every test
The 'should announce status changes using two alternately updated aria live regions' test relies on the autocomplete being in a particular state – the `bump` boolean needs to be false so that the first status ends up in the '__status--A` element rather than `__status--B`. The `bump` boolean is flipped every time a status update is made.

Depending on the number of status updates that have been triggered by previous tests, this means that we cannot guarantee that this will be the case. Because this boolean is encapsulated within the component's React state, we can neither observe nor modify its value.

This problem has been masked by the use of `retries` – if it fails the first time, it should pass the next time as it makes 3 status updates, so the `bump` state should be negated when it's re-tried.

The easiest way to work around this is to reload the page between every test – thus the component should be in the same state as when it was initialised.

We looked at using `browser.reload()` for this, but that is confusingly named as it actually reloads the entire Selenium session, and 'forgets' the URL that is being visited. Calling `browser.url` before each test does what we need.

Co-authored-by: Hanna Laakso <hanna.laakso@digital.cabinet-office.gov.uk>