Pin Sauce Connect Proxy version
v4.6.1, released May 19 2020, disables the Selenium Relay feature by default[1]. The current version of wdio-sauce-service relies on this feature in order to proxy requests from SauceLabs to the server running on localhost in the Travis VM.

Although it can be enabled on a specified port using the --se-port option, there doesn't appear to be a way to configure wdio-sauce-service or sauce-connect-launcher to do this.

Pin the version of Sauce Connect Proxy to the immediately preceding version (4.5.4) until we can do the work to upgrade to the latest version of Webdriver.io / wdio-sauce-service which does not rely on Selenium Relay.

[1]: https://wiki.saucelabs.com/display/DOCS/Sauce+Connect+Proxy+Version+4.6.1