Use legacy JSON Wire Protocol with Internet Explorer 9
I cannot get IE9 to work on Sauce Labs with W3C capabilities. I think
this is due to an issue in their stack. For now we can use the JSON Wire
Protocol (requires @wdio/sauce-service v6.3.0 or greater), maybe at a
future date Sauce Labs will fix the issue.

Detailed notes follow:

When using the W3C protocol with IE9 Sauce Labs adds the capability
`selenium:webdriver.remote.quietExceptions` to the session for IE9 (but
not IE10). This capability is not recognised by IE, and something in the
Sauce Labs stack complains about this. See the [test logs from Sauce
Labs] for more details about this.

I believe this is related to this [issue in IEDriverServer], but I could
be wrong. It's also not clear to me why Sauce Labs is adding this
capability for IE9 but not IE10, so Sauce Labs may change that behaviour
at some point.

As to why IE9 was working with WebdriverIO v4, I think that Sauce Labs
uses a different (legacy) stack if the client uses the JSON Wire
Protocol, which WebdriverIO v4 does, instead of the W3C Webdriver
protocol. WebdriverIO v6.3.0 includes a [fix to the Sauce service to
allow using the JSON Wire Protocol].

I also tried downgrading the version of Selenium used by Sauce Labs by
setting ['sauce:options']['seleniumVersion'], however that failed
because it expects the older JSON Wire Protocl, and although the commands
WebdriverIO v6 are backwards compatible with JSON Wire Protocol, when
using the newer W3C stack SauceLabs strips out the JSON Wire Protocol
`desiredCapabilties' key that Selenium v2 requires.

For information about why Sauce Labs is adding this capability to the
session, see the [Sauce Labs test configuration options documentation].
I'm not sure that this documentation is up to date though; for instance,
it doesn't cover how to make your capabilities W3C compatible, and
includes information about the Selenium version which doesn't seem to be
followed any more.

[test logs from Sauce Labs]: https://app.saucelabs.com/tests/4a28429a66ee4db7bf159fc43fdd4db3#0
[issue in IEDriverServer]: https://github.com/SeleniumHQ/selenium/issues/7479
[fix to the Sauce service to allow using the JSON Wire Protocol]: https://github.com/webdriverio/webdriverio/pull/5610
[Sauce Labs test configuration options documenation]: https://wiki.saucelabs.com/display/DOCS/Test+Configuration+Options