Pin Sauce Connect Proxy version to fix local tests
https://github.com/alphagov/accessible-autocomplete/pull/443 pinned the Sauce
Connect Proxy for Travis. https://github.com/alphagov/accessible-autocomplete/pull/450
will update WebDriverIO v6 so we no longer need to pin the version - the PR is
currently blocked as we’re speaking to SauceLabs to see if we can enable the tests for
IE9 with WebDriverIO v6.

This PR adds a temporary fix to unblock https://github.com/alphagov/accessible-autocomplete/pull/447
- we can remove this once  https://github.com/alphagov/accessible-autocomplete/pull/450
is merged.