Update Travis config to stop Travis exceeding SauceLabs quota
I've noticed that GitHub checks for a Pull Request often fail because
SauceLabs only allows 5 connections at a time[[1]], but the Travis tests will
try and run the test suite twice (once as branch and once as a PR).

This commit changes the Travis config to split the tests into three
parts, with the tests that use SauceLabs running only for Pull Requests,
not for branches. Hopefully this should reduce the number of failures.

[1]: https://saucelabs.com/solutions/open-source