Update Travis to use Ubuntu Xenial
Our tests on Travis are currently failing to start, I think because
we're using Ubuntu Trusty and a lot of the old apt servers needed are
disappearing (see this recent failed build log [[1]] for example).

This commit should fix things by upgrading to Xenial, which is
the default for Travis. This also means we can simplify the setup of
xvfb, following the suggestion in the Travis migration guide [[2]].

1: https://app.travis-ci.com/github/alphagov/accessible-autocomplete/builds/239180081
2: https://docs.travis-ci.com/user/trusty-to-xenial-migration-guide