Add GitHub Actions workflow for tests
Adds two workflows, `test.yaml` and `saucelabs.yaml`.

test.yaml covers all tests except those that use Sauce Labs, which are
in the saucelabs.yaml file. They are separated this way because the
Sauce Labs tests are a bit more resource intensive, and we want to run
them only for pull requests (and with a concurrency limit).