Bump Node.js from 10.24.1 to 14.18.2
We'd like to use the latest Long Term Support version of Node, which is
version 16. Unfortunately the version of the test runner WebdriverIO we
are currently using (v6) doesn't support Node.js 16 [[1]], and the next
version of WebdriverIO doesn't support Node.js 10 [[2]], so we have to
settle for Node.js 14.

[1]: https://webdriver.io/blog/2021/07/28/sync-api-deprecation/
[2]: https://webdriver.io/blog/2021/02/09/webdriverio-v7-released/#drop-node-v10-support