Ensure highlighted option is distinguishable in forced color mode
In forced colours mode (for example, Windows High Contrast Mode) the background and foreground colours of elements are forced to the users preferred colour scheme.

This has the unfortunate side effect of making it impossible to determine which option is currently selected when using the keyboard to move between results in the list, or which item in the list is currently being hovered over.

We can work around this by using the `forced-colors: active` media query (combined with the proprietary `-ms-high-contrast: active` to support IE11) to detect that forced colour mode is active, and then set the background and foreground of the highlighted / hovered option to use the system Highlight and HighlightText colours respectively.

In Chromium based browsers, a backplate is also being added behind the text for each option. The backplate uses the user's preferred background colour, and ensures that text remains legible if it's drawn over other elements. However, there is no guarantee that there will be sufficent contrast between HighlightText and this background colour – only between HighlightText and Highlight – so we need to disable the backplates within these elements using `forced-color-adjust: none` otherwise they can obscure the text.