Improve Firefox forced colors mode
In Firefox 96 the use of the Highlight colour pair can result in some relatively low contrast background/foreground colour combinations, and there is also no guarantee that there will be sufficent contrast between the highlighted and non-highlighted options for the user to be able to disambiguate them.

The definitions for the Highlight and SelectedItem colour pairs [1] in the current latest draft of CSS Color Module Level 4 suggest that the SelectedItem colour pair is a better choice for highlighting the selected option:

Highlight
  Background of selected text, for example from ::selection.
HighlightText
  Text of selected text.
SelectedItem
  Background of selected items, for example a selected checkbox.
SelectedItemText
  Text of selected items.

This is reinforced by the fact that Firefox uses the SelectedItem colour pair for selected options within selects [2].

However, the SelectedItem colour pair was only recently introduced [3], and is not currently a recognised system colour in some browsers (including Chrome 96, Safari 15.2).

Re-declaring the colours allows us to use the SelectedItem colour pair in browsers that support it (which includes Firefox) but fall back to the Highlight colour pair for those that do not.

Using the Field colour pair for non-highlighted options makes it more likely that there will be sufficient contrast between the highlighted and non-highlighted options to be able to disambiguate them. This only seems to affect Firefox, where text inputs (and the Field colour pair) always seem to use black text on white when colours are overridden.

[1]: https://www.w3.org/TR/css-color-4/#css-system-colors
[2]: https://searchfox.org/mozilla-central/rev/f8576fec48d866c5f988baaf1fa8d2f8cce2a82f/layout/style/res/forms.css#357-361
[3]: https://github.com/w3c/csswg-drafts/issues/6008