Note issue with forced colors and custom templates
Because we've opted out of forced-color-adjust (in order to remove backplates – see 0b50d73), there is a potential issue that could arise if custom templates are being used that include a child element with a defined text colour.

Those child elements will no longer be forced to the user's preferred colour, nor will they be forced to HighlightText, and so we cannot guarantee there will be sufficient contrast between that text and the background.

We think this is an edge case – and not something that the component should defend against – but worth flagging in the documentation for custom templates.