Use devtools protocol instead of Selenium for local tests
For the integration tests, when we're not communicating with Sauce Labs
we don't need to use the WebDriver protocol, we can use the Chromium
devtools protocol, as we're only controlling Chromium. This commit
changes the Webdriver.io test runner config to use devtools (using
instructions from [[1]]) and Puppeteer when not using Sauce Labs.

[1]: https://webdriver.io/docs/automationProtocols