Fix tests when using Chrome DevTools
We have a test that I _think_ checks that, if you somehow manage to
click on the text rather than the list item containing the text, the
outcome should be the same.

However, it seems that ChromeDriver doesn't allow clicking on span
elements [[1]], and throws an error. We previously worked around this by
having logic that either expected the click to succeed, or expected an
exception, depending on which browser was being tested.

However, because whether the error is thrown depends on the browser
automation protocol rather than the browser itself, and we're now using
the DevTools protocol, testing against Chrome will raise an exception
when communicating with Sauce Labs, but not when the tests are running
locally.

This commit changes the logic so that instead of checking the browser or
automation protocol (which is finicky and hard work), we just wrap the
click in a try catch and skip the rest of the test if there is an error.
In a sense the error 'Other element would receive the click' is good
enough to assert that the correct behaviour would be followed (assuming
the other element is the <li> as desired).

I don't know enough about frontend to know whether this test is actually
needed; based on what I think it's doing, it does seem like its testing
normal browser behaviour that you should be able to rely on, but I guess
it's possible to break things in ways that I'm not aware of... so I left
it in :D

[1]: https://stackoverflow.com/questions/11908249/debugging-element-is-not-clickable-at-point-error#comment48848194_19763087