Don't show tNoResults when debouncing
When a `source` query has yet to be called, it's not necessary to show
"no results found". Pass the debouncing state of the autocomplete down
to the Status component through Status.props.autocompleteDebouncing
such that the status component does not announce "no results available"
immediately before announcing "50 results available".

From the autocomplete, this is difficult to test. Tests already exist
showing that `this.state.debouncing` is set appropriately when
`debounceMs` is in use. It is used in conjunction with the
`showNoOptionsFound` prop to suppress tNoResults in render().

No good option exists for testing it that doesn't involve extra
redundant state to service the tests. It can't be done by rendering
tests alone, and it can't be done by behaviour tests alone.