Add accessibility controls to 'No results' list item
Adds `role="option"` and `aria-disabled="true"` to the rendered li when no results have been found. This ensures for screen readers who navigate to the 'no results' li that they know it's still semantically part of the autocomplete interface but not a valid option.

This solution was recommended by DAC following a recent audit of the GOV.UK Design System website.